refactor(events): rename EventAnnouncementEnBiweekly -> EnBiweeklyAnnouncement
